Create codeArduino.ino
First Arduino Code.
Inputs the data via Terminal, stores the into one of 3 spaces.
Two times the same data in more than one space is not allowed.
Is able to detect if spaces are full and all individual spaces can be made free with a simple command.

Servs as starting point.
WIFI needs to be added as well as the MQTT connection and chanels.